requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
ariadne-codegen==0.14.0
attrs==25.3.0
autoflake==2.3.1
backoff==2.2.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
frozenlist==1.5.0
gql==3.5.2
graphql-core==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
isort==6.0.1
multidict==6.2.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pip==25.0
platformdirs==4.3.7
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.8.0
sgqlc==16.3
six==1.17.0
sniffio==1.3.1
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
yarl==1.18.3
